Refactor, extract Ui class
